split up init.py and move things into pdf.py
